refactor(staking): reorganize contract implementation
