+++ .gitignore
@@ -34,6 +34,7 @@ contracts/target
 examples/target
 
 genesis-*.json
-node_modules
+node_modules/
 evm-e2e/tests/
-datadir
+datadir/
+graphify-out/
