+++ contracts/staking/src/lib.rs
@@ -12,13 +12,10 @@ extern crate alloc;
 mod config;
 mod consensus;
 mod consts;
-mod economics;
-mod equivocation;
 mod events;
-mod handlers;
-mod liveness;
+mod initializer;
 mod math;
-mod rewards;
+mod staking;
 mod storage;
 mod types;
 mod util;
@@ -42,28 +39,24 @@ pub fn main_entry<SDK: SharedAPI>(sdk: &mut SDK) -> Result<(), ExitCode> {
             .try_into()
             .map_err(|_| ExitCode::MalformedBuiltinParams)?,
     );
-    if let Some(result) = config::dispatch_constant(sdk, selector) {
-        return result;
-    }
     match selector {
-        // Deployment and governance configuration.
-        SIG_INITIALIZE => handlers::initialize(sdk, params),
-        SIG_CONFIGURE => handlers::configure(sdk, params),
-        SIG_CONFIGURE_DEPENDENCIES => config::configure_dependencies(sdk, params),
-        // Compatibility getters and governance parameter access.
-        SIG_CURRENT_EPOCH => handlers::current_epoch_read(sdk),
-        SIG_NEXT_EPOCH => handlers::next_epoch_read(sdk),
-        SIG_OWNER => handlers::owner(sdk),
-        SIG_GET_STAKING => handlers::get_staking(sdk),
-        SIG_GET_CHAIN_CONFIG => handlers::get_chain_config(sdk),
-        SIG_GET_GOVERNANCE => handlers::get_governance(sdk),
-        SIG_GET_STAKING_TOKEN => handlers::get_staking_token(sdk),
-        SIG_GET_ACTIVE_VALIDATORS_LENGTH => handlers::get_active_validators_length(sdk),
-        SIG_GET_EPOCH_BLOCK_INTERVAL => handlers::get_epoch_block_interval(sdk),
-        SIG_GET_DPOS_ACTIVATION_BLOCK => handlers::get_dpos_activation_block(sdk),
-        SIG_GET_UNDELEGATE_PERIOD => handlers::get_undelegate_period(sdk),
-        SIG_GET_MIN_VALIDATOR_STAKE_AMOUNT => handlers::get_min_validator_stake_amount(sdk),
-        SIG_GET_MIN_STAKING_AMOUNT => handlers::get_min_staking_amount(sdk),
+        // Initializer
+        SIG_INITIALIZE => initializer::initialize(sdk, params),
+
+        // ChainConfig
+        SIG_DEFAULT_PARTICIPATION_FLOOR_BPS => config::default_participation_floor_bps(sdk),
+        SIG_DEFAULT_SLASH_REPORTER_BPS => config::default_slash_reporter_bps(sdk),
+        SIG_MAX_ACTIVE_VALIDATORS => config::max_active_validators(sdk),
+        SIG_MAX_BLEND_STIPEND_PER_EPOCH => config::max_blend_stipend_per_epoch(sdk),
+        SIG_MAX_PARTICIPATION_FLOOR_BPS => config::max_participation_floor_bps(sdk),
+        SIG_MAX_SLASH_REPORTER_BPS => config::max_slash_reporter_bps(sdk),
+        SIG_GET_STAKING_TOKEN => config::get_staking_token(sdk),
+        SIG_GET_ACTIVE_VALIDATORS_LENGTH => config::get_active_validators_length(sdk),
+        SIG_GET_EPOCH_BLOCK_INTERVAL => config::get_epoch_block_interval(sdk),
+        SIG_GET_DPOS_ACTIVATION_BLOCK => config::get_dpos_activation_block(sdk),
+        SIG_GET_UNDELEGATE_PERIOD => config::get_undelegate_period(sdk),
+        SIG_GET_MIN_VALIDATOR_STAKE_AMOUNT => config::get_min_validator_stake_amount(sdk),
+        SIG_GET_MIN_STAKING_AMOUNT => config::get_min_staking_amount(sdk),
         SIG_GET_FELONY_THRESHOLD => config::get_felony_threshold(sdk),
         SIG_SET_FELONY_THRESHOLD => config::set_felony_threshold(sdk, params),
         SIG_GET_VALIDATOR_JAIL_EPOCH_LENGTH => config::get_validator_jail_epoch_length(sdk),
@@ -88,30 +81,44 @@ pub fn main_entry<SDK: SharedAPI>(sdk: &mut SDK) -> Result<(), ExitCode> {
         SIG_SET_BLS_VERIFIER => config::set_bls_verifier(sdk, params),
         SIG_GET_EVIDENCE_DECODER => config::get_evidence_decoder(sdk),
         SIG_SET_EVIDENCE_DECODER => config::set_evidence_decoder(sdk, params),
-        // Validator stake and delegation principal.
-        SIG_GET_VALIDATOR_DELEGATION => economics::get_validator_delegation(sdk, params),
+
+        // Staking
+        SIG_CURRENT_EPOCH => staking::current_epoch_read(sdk),
+        SIG_NEXT_EPOCH => staking::next_epoch_read(sdk),
+        SIG_OWNER => staking::owner(sdk),
+        SIG_IS_VALIDATOR => staking::is_validator(sdk, params),
+        SIG_IS_VALIDATOR_ACTIVE => staking::is_validator_active(sdk, params),
+        SIG_GET_VALIDATOR_STATUS => staking::get_validator_status(sdk, params),
+        SIG_GET_VALIDATOR_BY_OWNER => staking::get_validator_by_owner(sdk, params),
+        SIG_GET_VALIDATORS => staking::get_validators(sdk),
+        SIG_ADD_VALIDATOR => staking::add_validator(sdk, params),
+        SIG_ACTIVATE_VALIDATOR => staking::activate_validator(sdk, params),
+        SIG_DISABLE_VALIDATOR => staking::disable_validator(sdk, params),
+        SIG_CHANGE_VALIDATOR_COMMISSION_RATE => staking::change_commission(sdk, params),
+        SIG_CHANGE_VALIDATOR_OWNER => staking::change_owner(sdk, params),
+        SIG_GET_VALIDATOR_DELEGATION => staking::get_validator_delegation(sdk, params),
         SIG_GET_VALIDATOR_DELEGATED_STAKE_AT => {
-            economics::get_validator_delegated_stake_at(sdk, params)
+            staking::get_validator_delegated_stake_at(sdk, params)
         }
-        SIG_REGISTER_VALIDATOR => economics::register_validator(sdk, params),
-        SIG_DELEGATE => economics::delegate(sdk, params),
-        SIG_UNDELEGATE => economics::undelegate(sdk, params),
-        // Validator and delegator rewards.
-        SIG_GET_VALIDATOR_FEE => rewards::get_validator_fee(sdk, params),
-        SIG_GET_PENDING_VALIDATOR_FEE => rewards::get_pending_validator_fee(sdk, params),
-        SIG_CLAIM_VALIDATOR_FEE => rewards::claim_validator_fee(sdk, params),
-        SIG_CLAIM_VALIDATOR_FEE_AT_EPOCH => rewards::claim_validator_fee_at_epoch(sdk, params),
-        SIG_GET_DELEGATOR_FEE => rewards::get_delegator_fee(sdk, params),
-        SIG_GET_PENDING_DELEGATOR_FEE => rewards::get_pending_delegator_fee(sdk, params),
-        SIG_CLAIM_DELEGATOR_FEE => rewards::claim_delegator_fee(sdk, params),
-        SIG_CLAIM_DELEGATOR_FEE_AT_EPOCH => rewards::claim_delegator_fee_at_epoch(sdk, params),
+        SIG_REGISTER_VALIDATOR => staking::register_validator(sdk, params),
+        SIG_DELEGATE => staking::delegate(sdk, params),
+        SIG_UNDELEGATE => staking::undelegate(sdk, params),
+        SIG_GET_VALIDATOR_FEE => staking::get_validator_fee(sdk, params),
+        SIG_GET_PENDING_VALIDATOR_FEE => staking::get_pending_validator_fee(sdk, params),
+        SIG_CLAIM_VALIDATOR_FEE => staking::claim_validator_fee(sdk, params),
+        SIG_CLAIM_VALIDATOR_FEE_AT_EPOCH => staking::claim_validator_fee_at_epoch(sdk, params),
+        SIG_GET_DELEGATOR_FEE => staking::get_delegator_fee(sdk, params),
+        SIG_GET_PENDING_DELEGATOR_FEE => staking::get_pending_delegator_fee(sdk, params),
+        SIG_CLAIM_DELEGATOR_FEE => staking::claim_delegator_fee(sdk, params),
+        SIG_CLAIM_DELEGATOR_FEE_AT_EPOCH => staking::claim_delegator_fee_at_epoch(sdk, params),
         SIG_CALC_AVAILABLE_FOR_REDELEGATE_AMOUNT => {
-            rewards::calc_available_for_redelegate_amount(sdk, params)
+            staking::calc_available_for_redelegate_amount(sdk, params)
         }
-        SIG_REDELEGATE_DELEGATOR_FEE => rewards::redelegate_delegator_fee(sdk, params),
-        SIG_GET_EPOCH_REWARDS => rewards::get_epoch_rewards(sdk, params),
-        SIG_SETTLE_EPOCH_STIPEND => rewards::settle_epoch_stipend(sdk, params),
-        // Consensus identities and committed committees.
+        SIG_REDELEGATE_DELEGATOR_FEE => staking::redelegate_delegator_fee(sdk, params),
+        SIG_GET_EPOCH_REWARDS => staking::get_epoch_rewards(sdk, params),
+        SIG_SETTLE_EPOCH_STIPEND => staking::settle_epoch_stipend(sdk, params),
+
+        // Consensus
         SIG_SET_CONSENSUS_KEYS => consensus::set_consensus_keys(sdk, params),
         SIG_GET_CONSENSUS_KEYS => consensus::get_consensus_keys(sdk, params),
         SIG_GET_VALIDATORS_WITH_KEYS => consensus::get_validators_with_keys(sdk),
@@ -127,31 +134,18 @@ pub fn main_entry<SDK: SharedAPI>(sdk: &mut SDK) -> Result<(), ExitCode> {
         SIG_GET_EPOCH_COMMITTEE_WITH_STAKES => {
             consensus::get_epoch_committee_with_stakes(sdk, params)
         }
-        // Liveness and equivocation penalties.
-        SIG_RELEASE_VALIDATOR_FROM_JAIL => liveness::release_validator_from_jail(sdk, params),
-        SIG_READMIT_EXPIRED_JAILS => liveness::readmit_expired_jails(sdk, params),
-        SIG_SLASH => liveness::slash(sdk, params),
-        SIG_COMMIT_EQUIVOCATION_REPORT => equivocation::commit_report(sdk, params),
+        SIG_RELEASE_VALIDATOR_FROM_JAIL => consensus::release_validator_from_jail(sdk, params),
+        SIG_READMIT_EXPIRED_JAILS => consensus::readmit_expired_jails(sdk, params),
+        SIG_SLASH => consensus::slash(sdk, params),
+        SIG_COMMIT_EQUIVOCATION_REPORT => consensus::commit_report(sdk, params),
         SIG_COMPUTE_EQUIVOCATION_REPORT_COMMITMENT => {
-            equivocation::compute_report_commitment(sdk, params)
+            consensus::compute_report_commitment(sdk, params)
         }
-        SIG_GET_EQUIVOCATION_REPORT_COMMITMENT => equivocation::get_report_commitment(sdk, params),
-        SIG_SLASH_EQUIVOCATION_NOTARIZE => equivocation::slash_notarize(sdk, params),
-        SIG_SLASH_EQUIVOCATION_FINALIZE => equivocation::slash_finalize(sdk, params),
-        SIG_SLASH_EQUIVOCATION_NULLIFY_FINALIZE => {
-            equivocation::slash_nullify_finalize(sdk, params)
-        }
-        // Validator lifecycle and ownership.
-        SIG_IS_VALIDATOR => handlers::is_validator(sdk, params),
-        SIG_IS_VALIDATOR_ACTIVE => handlers::is_validator_active(sdk, params),
-        SIG_GET_VALIDATOR_STATUS => handlers::get_validator_status(sdk, params),
-        SIG_GET_VALIDATOR_BY_OWNER => handlers::get_validator_by_owner(sdk, params),
-        SIG_GET_VALIDATORS => handlers::get_validators(sdk),
-        SIG_ADD_VALIDATOR => handlers::add_validator(sdk, params),
-        SIG_ACTIVATE_VALIDATOR => handlers::activate_validator(sdk, params),
-        SIG_DISABLE_VALIDATOR => handlers::disable_validator(sdk, params),
-        SIG_CHANGE_VALIDATOR_COMMISSION_RATE => handlers::change_commission(sdk, params),
-        SIG_CHANGE_VALIDATOR_OWNER => handlers::change_owner(sdk, params),
+        SIG_GET_EQUIVOCATION_REPORT_COMMITMENT => consensus::get_report_commitment(sdk, params),
+        SIG_SLASH_EQUIVOCATION_NOTARIZE => consensus::slash_notarize(sdk, params),
+        SIG_SLASH_EQUIVOCATION_FINALIZE => consensus::slash_finalize(sdk, params),
+        SIG_SLASH_EQUIVOCATION_NULLIFY_FINALIZE => consensus::slash_nullify_finalize(sdk, params),
+
         _ => revert(sdk, ERR_UNKNOWN_METHOD),
     }
 }
